polygon + fixing clear screen freeze bug

+++ core/point.h
@@ -24,7 +24,7 @@ public:
     }
 
     double slope(const Point& other) const {
-        return (double)dy(other) / (double)dx(other);
+        return dy(other) / dx(other);
     }
 
     // Distance
